feat(nexus): add graph network view with bucket resolution and filters
- Added new GraphNetworkView component for network visualization.
- BFO bucket resolution logic for instance-graph nodes in resolveGraphNodeBucket.ts.
- Integrated GraphNetworkView in graph page replacing prior inline implementation.
- Added UI controls for node search, relations toggle, parents hierarchy expansion.
- Improved graph parent nodes fetching and cache management.

Test plan:
- Verify new graph network view loads with nodes and edges.
- Check filters and parent expansion UI functionality.
- Confirm integration in graph page does not regress other views.
